requirements.txt

openai
sendgrid
python-dotenv
certifi
